Partial (not working) impl of hot reloadable dll...
Turns out DLL code doesn't seem to have access to the Raylib created
opengl context and so can't perform gl related calls because the context
is null in that... context.

Going to retool so that all raylib and gl context dependent work happens
in the dll code rather than in the main executable code.